test: move vaultActionsLogic mock and spy to same file as vaultActionsLogic
